Update route and move controller in customer folder
